Add prompt to deploy to ask about unit tests

+++ deploy.sh
@@ -1,4 +1,14 @@
-#!/bin/sh -x
+#!/usr/bin/env bash
+
+read -p "Did you run all unit tests? [y/n] " -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+  echo "Stopping so you can run unit tests and make sure they all pass."
+  exit 0
+fi
+
+set -x
+
 rm -rf /var/www/html/materialscommons.old
 mv /var/www/html/materialscommons /var/www/html/materialscommons.old
 cd ..
